mv xpose config gen to utils
